[9/20 19:03 lx]ID的数据类型改为Long

+++ fund-service-api/src/main/java/com/example/fundserviceapi/client/FundClient.java
@@ -8,16 +8,16 @@ import org.springframework.web.bind.annotation.RequestParam;
 
 @FeignClient("fund-service")
 public interface FundClient {
-    //采购单消息     采购单cgdid/客户cstid/结算账户accid/金额account
+    //采购单消息     采购单cgdid/供应商gysi/结算账户accid/金额account
     @PostMapping("/billpay/addBillpayMsg")
-    public ResponseResult<Void> getCgdMsg(@RequestParam("cgdid")Integer cgdid,
-                                          @RequestParam("gysid")Integer gysid,
-                                          @RequestParam("accid")Integer accid,
+    public ResponseResult<Void> getCgdMsg(@RequestParam("cgdid")Long cgdid,
+                                          @RequestParam("gysid")Long gysid,
+                                          @RequestParam("accid")Long accid,
                                           @RequestParam("account")Double account);
-    //出货单消息     出货单chdid/供应商gysid/结算账户accid/金额account
+    //出货单消息     出货单chdid/d客户cstid/结算账户accid/金额account
     @PostMapping("/billrece/addBillreceMsg")
-    public ResponseResult<Void> getChdMsg(@RequestParam("chdid")Integer chdid,
-                                          @RequestParam("cstid")Integer cstid,
-                                          @RequestParam("accid")Integer accid,
+    public ResponseResult<Void> getChdMsg(@RequestParam("chdid")Long chdid,
+                                          @RequestParam("cstid")Long cstid,
+                                          @RequestParam("accid")Long accid,
                                           @RequestParam("account")Double account);
 }
